[Android] Move ServiceTreeItem back to model
