+++ src/android/com/microsoftopentechnologies/intellij/model/ms/Table.java
@@ -16,6 +16,8 @@
 
 package com.microsoftopentechnologies.intellij.model.ms;
 
+import com.microsoftopentechnologies.intellij.model.ServiceTreeItem;
+
 import java.util.ArrayList;
 
 public class Table implements ServiceTreeItem {
