+++ src/android/com/microsoftopentechnologies/intellij/model/vm/Endpoint.java
@@ -15,7 +15,7 @@
  */
 package com.microsoftopentechnologies.intellij.model.vm;
 
-import com.microsoftopentechnologies.intellij.model.ms.ServiceTreeItem;
+import com.microsoftopentechnologies.intellij.model.ServiceTreeItem;
 import org.jetbrains.annotations.NotNull;
 
 public class Endpoint implements ServiceTreeItem {
